INTERLOK-2792 Refactored the tests, seeing as they are exactly the same for both PdfTo..X services.  Added a new PdfToText Service.
